Move models to a models folder
